+++ web/src/pages/home/Accommodations.js
@@ -5,7 +5,7 @@ import PageSwitcher from '../../utils/PageSwitcher.module.css';
 import SkeletonLoader from '../../components/base/SkeletonLoader';
 import { useNavigate } from 'react-router-dom';
 import AccommodationCard from "./AccommodationCard";
-import FiltersMain from "./FiltersMain";
+import FilterUi from "./FilterUi";
 
 const Accommodations = ({ searchResults }) => {
   const [accolist, setAccolist] = useState([]);
@@ -107,7 +107,7 @@ const Accommodations = ({ searchResults }) => {
   return (
     <div id="container">
       <div id="filters-sidebar">
-        <FiltersMain />
+        <FilterUi />
       </div>
       <div id="card-visibility">
         {displayedAccolist.length > 0 ? (
