+++ web/src/pages/home/FilterUi.jsx
@@ -1,88 +1,32 @@
-// code gemaakt door: @kacperfl 
-//importeren van de benodigde modules
-import React, { useState } from 'react';
+// FilterUi.jsx
+import React from 'react';
 import Slider from '@mui/material/Slider';
+import { FilterLogic } from './FilterLogic';
 import './FilterMain.css';
 
-const FilterModal = () => {
-  const [priceValues, setPriceValues] = useState([15, 400]);
-  const [selectedFacilities, setSelectedFacilities] = useState({
-    // Alle mogelijke faciliteiten die een accommodatie kan hebben
-    wifi: false,
-    parking: false,
-    gym: false,
-    spa: false,
-    swimmingPool: false,
-    restaurant: false,
-    petFriendly: false,
-    airConditioning: false,
-    breakfast: false,
-    bar: false,
-  });
-// Alle mogelijke soorten accommodaties die een accommodatie kan hebben
-  const [selectedPropertyTypes, setSelectedPropertyTypes] = useState({
-    apartment: false,
-    villa: false,
-    bungalow: false,
-    studio: false,
-    penthouse: false,
-    cottage: false,
-    townhouse: false,
-    cabin: false,
-    chalet: false,
-    duplex: false,
-    mansion: false,
-    farmstay: false,
-  });
-  //op default staat de optie voor tonen van meer opties op false
-  const [showMoreFacilities, setShowMoreFacilities] = useState(false);
-  const [showMorePropertyTypes, setShowMorePropertyTypes] = useState(false);
-
-  //Zorgt ervoor dat de ingevoerde prijs waarden correct worden bijgewerkt maar voorkomt dat de minimumprijs hoger wordt dan de maximumprijs.
-  const handleInputChange = (index, value) => {
-    const newValues = [...priceValues];
-    newValues[index] = Number(value);
-    if (newValues[0] <= newValues[1]) {
-      setPriceValues(newValues);
-    }
-  };
-//Functie om de geselecteerde faciliteiten aan te passen
-  const handleFacilityChange = (event) => {
-    setSelectedFacilities({
-      ...selectedFacilities,
-      [event.target.name]: event.target.checked,
-    });
-  };
-//Functie om de geselecteerde soorten accommodaties aan te passen
-  const handlePropertyTypeChange = (event) => {
-    setSelectedPropertyTypes({
-      ...selectedPropertyTypes,
-      [event.target.name]: event.target.checked,
-    });
-  };
-//Functie dat de checkoxes voor de ratings op default uit zet
-  const [selectedRatings, setSelectedRatings] = useState({
-    1: false,
-    2: false,
-    3: false,
-    4: false,
-    5: false,
-  });
-//Functie om de geselecteerde ratings aan te passen
-  const handleRatingChange = (event) => {
-    setSelectedRatings({
-      ...selectedRatings,
-      [event.target.name]: event.target.checked,
-    });
-  };
+const FilterUi = () => {
+  const {
+    priceValues,
+    setPriceValues,
+    selectedFacilities,
+    handleFacilityChange,
+    selectedPropertyTypes,
+    handlePropertyTypeChange,
+    showMoreFacilities,
+    setShowMoreFacilities,
+    showMorePropertyTypes,
+    setShowMorePropertyTypes,
+    selectedRatings,
+    handleRatingChange,
+    handlePriceChange,
+  } = FilterLogic();
 
   return (
     <div>
-      {/* Prijsfilter met een slider*/}
+      {/* Price filter with slider */}
       <div className="filter-section">
         <div className='FilterTitle'>Price Range</div>
         <div className="slider-container">
-          {/* Slider voor het selecteren van de prijs met styling er in */}
           <Slider
             sx={{
               '& .MuiSlider-thumb': {
@@ -102,7 +46,6 @@ const FilterModal = () => {
                 backgroundColor: '#e2e2e2',
               },
             }}
-            //waarden voor de slider en de functie die de waarden bijwerkt
             value={priceValues}
             onChange={(e, newValues) => setPriceValues(newValues)}
             valueLabelDisplay="auto"
@@ -114,26 +57,24 @@ const FilterModal = () => {
           />
           <div className="price-inputs">
             <div>
-              {/* Input veld voor de minimale prijs */}
               <label>Min:</label>
               <input
                 type="text"
                 value={`€${priceValues[0] || ''}`}
                 onChange={(e) => {
                   const value = e.target.value.replace(/[^0-9]/g, '');
-                  handleInputChange(0, value);
+                  handlePriceChange(0, value);
                 }}
               />
             </div>
             <div>
-              {/* Input veld voor de maximale prijs */}
               <label>Max:</label>
               <input
                 type="text"
                 value={`€${priceValues[1] || ''}`}
                 onChange={(e) => {
                   const value = e.target.value.replace(/[^0-9]/g, '');
-                  handleInputChange(1, value);
+                  handlePriceChange(1, value);
                 }}
               />
             </div>
@@ -141,12 +82,10 @@ const FilterModal = () => {
         </div>
       </div>
 
-      {/* Facility filter met checkboxes */}
+      {/* Facility filter with checkboxes */}
       <div className="filter-section">
         <div className='FilterTitle'>Facilities</div>
         <div className="facility-list">
-
-          {/* Alle mogelijke faciliteiten die een accommodatie kan hebben */}
           {Object.keys(selectedFacilities).slice(0, 5).map((facility) => (
             <label key={facility} className="facility-item">
               <input
@@ -159,7 +98,6 @@ const FilterModal = () => {
               {facility.charAt(0).toUpperCase() + facility.slice(1)}
             </label>
           ))}
-          {/* Als er meer dan 5 faciliteiten zijn, dan worden de overige faciliteiten verborgen en kan de gebruiker deze tonen door op de tekst te klikken */}
           {showMoreFacilities &&
             Object.keys(selectedFacilities)
               .slice(5)
@@ -176,7 +114,6 @@ const FilterModal = () => {
                 </label>
               ))}
         </div>
-        {/* Functie om meer opties aan gebruiker te laten tonen */}
         <span
           onClick={() => setShowMoreFacilities(!showMoreFacilities)}
           className="show-more-text"
@@ -185,11 +122,10 @@ const FilterModal = () => {
         </span>
       </div>
 
-      {/* Property type filter met checkboxes */}
+      {/* Property type filter with checkboxes */}
       <div className="filter-section">
         <div className='FilterTitle'>Property Type</div>
         <div className="facility-list">
-          {/* Alle mogelijke soorten accommodaties die een accommodatie kan hebben */}
           {Object.keys(selectedPropertyTypes).slice(0, 5).map((propertyType) => (
             <label key={propertyType} className="facility-item">
               <input
@@ -202,7 +138,6 @@ const FilterModal = () => {
               {propertyType.charAt(0).toUpperCase() + propertyType.slice(1)}
             </label>
           ))}
-           {/* Als er meer dan 5 soorten accommodaties zijn, dan worden de overige soorten accommodaties verborgen en kan de gebruiker deze tonen door op de tekst te klikken */}
           {showMorePropertyTypes &&
             Object.keys(selectedPropertyTypes)
               .slice(5)
@@ -219,7 +154,6 @@ const FilterModal = () => {
                 </label>
               ))}
         </div>
-        {/* Functie om meer opties aan gebruiker te laten tonen */}
         <span
           onClick={() => setShowMorePropertyTypes(!showMorePropertyTypes)}
           className="show-more-text"
@@ -228,7 +162,7 @@ const FilterModal = () => {
         </span>
       </div>
 
-      {/* Rating filter met checkboxes */}
+      {/* Rating filter with checkboxes */}
       <div className="filter-section">
         <div className='FilterTitle'>Property Rating</div>
         <div className="facility-list">
@@ -250,4 +184,4 @@ const FilterModal = () => {
   );
 };
 
-export default FilterModal;
+export default FilterUi;
